[C#] refactor: use fluent api in samples
