+++ dotnet/samples/04.e.twentyQuestions/TwentyQuestions.csproj
@@ -12,7 +12,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.19.3" />
+    <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.21.0" />
     <PackageReference Include="Microsoft.TeamsAI" Version="1.0.*-*" />
   </ItemGroup>
